refactor: separate system toml config from service management logic
