+++ crates/common/tedge_config/src/system_services/manager.rs
@@ -1,6 +1,7 @@
 use camino::Utf8Path;
 
-use crate::system_services::*;
+use super::*;
+use crate::SystemTomlError;
 use std::fmt::Debug;
 use std::sync::Arc;
 
@@ -34,6 +35,6 @@ pub trait SystemServiceManager: Debug {
 
 pub fn service_manager(
     config_root: &Utf8Path,
-) -> Result<Arc<dyn SystemServiceManager>, SystemServiceError> {
+) -> Result<Arc<dyn SystemServiceManager>, SystemTomlError> {
     Ok(Arc::new(GeneralServiceManager::try_new(config_root)?))
 }
