+++ crates/common/tedge_config/src/system_services/managers/mod.rs
@@ -1,5 +1,3 @@
-mod config;
 mod general_manager;
 
-pub use self::config::*;
 pub use self::general_manager::*;
