+++ crates/common/tedge_config/src/system_services/mod.rs
@@ -10,7 +10,6 @@
 
 mod command_builder;
 mod error;
-mod log_config;
 mod manager;
 mod manager_ext;
 mod managers;
@@ -18,7 +17,6 @@ mod services;
 
 pub use self::command_builder::*;
 pub use self::error::*;
-pub use self::log_config::*;
 pub use self::manager::*;
 pub use self::manager_ext::*;
 pub use self::managers::*;
