+++ crates/core/tedge/src/main.rs
@@ -24,7 +24,7 @@ use tedge::TEdgeOpt;
 use tedge::TEdgeOptMulticall;
 use tedge_apt_plugin::AptCli;
 use tedge_config::cli::CommonArgs;
-use tedge_config::system_services::log_init;
+use tedge_config::log_init;
 use tracing::log;
 
 #[global_allocator]
