+++ crates/core/tedge_agent/src/restart_manager/actor.rs
@@ -15,8 +15,8 @@ use tedge_actors::Sender;
 use tedge_actors::SimpleMessageBox;
 use tedge_api::commands::CommandStatus;
 use tedge_api::RestartCommand;
-use tedge_config::system_services::SystemConfig;
-use tedge_config::system_services::SystemSpecificCommands;
+use tedge_config::SystemConfig;
+use tedge_config::SystemSpecificCommands;
 use tokio::process::Command;
 use tokio::time::timeout;
 use tracing::error;
